Create README.md
A fun game! A game that generates a random number and ask the player to guess it. If the player guess is higher than the actual the the game displays 'Lower number please'. Similarly, if the player guess is lower than the actual number than the game displays 'Higher number please'. Whenever the player guesses the correct number the program displays the number of guesses to approach the actual number. The player who guesses the correct number in minimum no of times will win the game.